backend/requirements.txt
fastapi
pymongo
python-dotenv
sqlalchemy
psycopg2-binary
pgvector
sentence-transformers
uvicorn
